server vs db
