+++ package.json
@@ -1,6 +1,7 @@
 {
     "name": "mongoapp",
     "version": "1.0.0",
+    "type": "module",
     "dependencies": {
         "body-parser": "^1.19.1",
         "express": "^4.17.0",
